fix open/closing shop bug + added shop music & deny audio + changed unpurchasable interaction
clicking exit button when in shop would not toggle mouse correctly.
added shop music to audiomanager because of bug & added deny audio that plays when player clicks the purchase button but not enough money
instead of deactivating the buy button when player doesn't have enough money, it is change to play a deny sound indicating that the player couldn't afford it.